Move test cpp to test directory
