+++ moveit_ros/trajectory_cache/src/trajectory_cache.cpp
@@ -22,7 +22,7 @@
 
 #include "rclcpp/rclcpp.hpp"
 #include "rclcpp/logging.hpp"
-#include "trajectory_cache.hpp"
+#include "moveit/trajectory_cache/trajectory_cache.hpp"
 
 namespace moveit_ros
 {
